centralize fuzzing helpers; run some basic quickcheck end-to-end fuzz tests on all CI targets
