+++ test-libz-rs-sys/Cargo.toml
@@ -10,14 +10,20 @@ homepage.workspace = true
 publish.workspace = true
 rust-version.workspace = true
 
+
+[features]
+default = ["rust-allocator"]
+c-allocator = ["libz-rs-sys/c-allocator"]
+rust-allocator = ["libz-rs-sys/rust-allocator"]
+
+[dependencies]
+zlib-rs = { workspace = true, default-features = false, features = ["std", "__internal-test"] }
+libz-rs-sys = { workspace = true, default-features = false, features = ["std", "testing-prefix"] }
+libz-sys.workspace = true
+
 [dev-dependencies]
 quickcheck.workspace = true
 crc32fast = "1.3.2"
 
-zlib-rs = { workspace = true, default-features = false, features = ["std", "__internal-test"] }
-libz-rs-sys = { workspace = true, default-features = false, features = ["std", "rust-allocator", "testing-prefix"] }
-
 libloading.workspace = true
 dynamic-libz-sys.workspace = true
-
-libz-sys.workspace = true
